+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Patch Notes
 
+## Version 3.4.0
+
+- Add initial experimental implementation of scene control in a webviewer scene.
+
+## Version 3.3.1
+
+- Add forgotten import.
+
 ## Version 3.3.0
 
 - Add support for scene webviewers.